backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pypdf
pydantic
pypdfium2
pillow
supabase